ref(stream): Create StreamContainer to manage data required for the stream
Move tests to correct folder
